refactor: server-mode queryables
